Refactor: Merge Network files
